Fixed possible bug with updating location and heading.
I have not managed to test this on a device yet, but I think there was
a bug here.

SCOcurrence instances start updating *heading* and *location* when
init'd. The delegate callback methods for updates to these values
stopped the updating, but for the other value.

`locationManager:didUpdateToLocation:fromLocation` was stopping the
heading updates, while `locationManager:didUpdateHeading` was stopping
location updates. I believe this would mean only one of the two would
ever be available, and it would be the first that got returned. It may
have also left the location manager updating the other value
indefinitely.